refactor special reducer into search reducer
